refactored code to separate workspace operations into a new module
