+++ src/preset/__main__.py
@@ -5,18 +5,19 @@ import shutil
 import subprocess
 
 from . import cli
+from . import workspace
 
 
 USERDIR = os.environ['USER_PWD']
 PYTHONBIN = os.environ['PYTHONBIN']
 
 
-def execute(workspace, extra_vars):
-    if os.path.exists(os.path.join(workspace, 'requirements.yml')):
+def execute(ws, extra_vars):
+    if os.path.exists(os.path.join(ws, 'requirements.yml')):
         galaxy = os.path.join(PYTHONBIN, 'ansible-galaxy')
         rc = subprocess.run(
             f'{galaxy} collection install -r requirements.yml',
-            cwd=workspace,
+            cwd=ws,
             shell=True,
             stdout=sys.stdout,
             stderr=sys.stderr,
@@ -25,14 +26,14 @@ def execute(workspace, extra_vars):
         if rc != 0:
             raise RuntimeError('Failed to install collection dependencies')
 
-    os.environ['ANSIBLE_INVENTORY'] = os.path.join(workspace, 'hosts')
+    os.environ['ANSIBLE_INVENTORY'] = os.path.join(ws, 'hosts')
 
     ansible = os.path.join(PYTHONBIN, 'ansible-playbook')
-    playbook = os.path.join(workspace, 'playbook.yml')
+    playbook = os.path.join(ws, 'playbook.yml')
 
     rc = subprocess.run(
         f'{ansible} {playbook} --extra-vars \'{extra_vars}\'',
-        cwd=workspace,
+        cwd=ws,
         env=os.environ,
         shell=True,
         stdout=sys.stdout,
@@ -44,39 +45,10 @@ def execute(workspace, extra_vars):
     return 0
 
 
-def clone(workspace, collection, playbook):
-    if os.path.exists(playbook):
-        shutil.copy2(playbook, os.path.join(workspace, 'playbook.yml'))
-    else:
-        raise FileNotFoundError(f'Playbook not found: {playbook}')
-
-    requirements = None
-    for r in ['requirements.yml', 'requirements.yaml']:
-        if os.path.exists(os.path.join(collection, r)):
-            requirements = os.path.join(collection, r)
-            break
-    if requirements:
-        shutil.copy2(requirements, os.path.join(workspace, 'requirements.yml'))
-
-    roles = os.path.join(collection, 'roles')
-    if os.path.exists(roles):
-        shutil.copytree(roles, os.path.join(workspace, 'roles'))
-
-    plugins = os.path.join(collection, 'plugins')
-    if os.path.exists(plugins):
-        shutil.copytree(plugins, os.path.join(workspace, 'plugins'))
-
-    os.makedirs(os.path.join(workspace, 'host_vars'), exist_ok=True)
-    with open(os.path.join(workspace, 'host_vars', 'localhost.yml'), 'w') as f:
-        f.write('---\n')
-        f.write('ansible_connection: local\n')
-        f.write(f'ansible_python_interpreter: {sys.executable}')
-
-    with open(os.path.join(workspace, 'hosts'), 'w') as f:
-        f.write('localhost')
 
 
-def process(workspace, preset):
+
+def process(ws, preset):
     if os.path.isabs(preset.collection):
         collection = preset.collection
     else:
@@ -99,14 +71,14 @@ def process(workspace, preset):
     if not os.access(playbook, os.R_OK):
         raise PermissionError(f'Playbook is not readable: {playbook}')
 
-    clone(workspace, collection, playbook)
-    return execute(workspace, preset.extra_vars)
+    workspace.clone(ws, collection, playbook)
+    return execute(ws, preset.extra_vars)
 
 
 def main():
     preset = cli.parse()
-    with tempfile.TemporaryDirectory(prefix='preset-') as workspace:
-        process(workspace, preset)
+    with tempfile.TemporaryDirectory(prefix='preset-') as ws:
+        process(ws, preset)
 
 if __name__ == '__main__':
     main()
