+++ src/preset/types.py
@@ -1,3 +1,5 @@
 import collections
 
 Preset = collections.namedtuple('Preset', ['collection', 'playbook', 'extra_vars'])
+
+Workspace = collections.namedtuple('Workspace', ['path', 'preset'])
